refactor(playback): split Warp / Loop Policy decisions into GhostPlaybackLogic.WarpLoopPolicy.cs
